Address review feedback: split parser, improve warnings, batch flush, extract helpers
